--- app/src/main/java/com/rempahpedia/rempahpedia/data/data/Spices.kt
@@ -1,10 +1,0 @@
-package com.rempahpedia.rempahpedia.data.data
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-@Parcelize
-data class Spices(
-    val name: String,
-    val img: Int
-) : Parcelable
-
